# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Faturação",
        "description": "Documentação automática gerada pelo Swagger para a API de Faturação.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/faturas": {
            "get": {
                "tags": [
                    "Faturas"
                ],
                "summary": "Lista todas as faturas",
                "operationId": "8b148ee68128220c5969c7198aeb8878",
                "responses": {
                    "200": {
                        "description": "Lista de faturas retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Faturas"
                ],
                "summary": "Cria uma nova fatura",
                "operationId": "18d8ee93ee1feb08768c8396662c1e91",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "data_emissao",
                                    "valor_total"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "data_emissao": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-01"
                                    },
                                    "valor_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fatura criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/faturas/{id}": {
            "get": {
                "tags": [
                    "Faturas"
                ],
                "summary": "Exibe uma fatura específica",
                "operationId": "cb743f4c0c9c0398cd7c6db69c6ea39e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fatura retornada com sucesso"
                    },
                    "404": {
                        "description": "Fatura não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Faturas"
                ],
                "summary": "Atualiza uma fatura existente",
                "operationId": "ef6aa2637d424d5522497b4b1b187981",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "data_emissao": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-01"
                                    },
                                    "valor_total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Fatura atualizada com sucesso"
                    },
                    "404": {
                        "description": "Fatura não encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Faturas"
                ],
                "summary": "Remove uma fatura",
                "operationId": "a2c402b1d4eb5bad2a2deda76bfbd700",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fatura removida com sucesso"
                    },
                    "404": {
                        "description": "Fatura não encontrada"
                    }
                }
            }
        },
        "/api/faturas/{id}/emitir": {
            "post": {
                "tags": [
                    "Faturas"
                ],
                "summary": "Emite uma fatura",
                "operationId": "f60c9f3cacd460577831cede7debcc91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fatura emitida com sucesso"
                    },
                    "404": {
                        "description": "Fatura não encontrada"
                    }
                }
            }
        },
        "/api/faturas/{id}/anular": {
            "post": {
                "tags": [
                    "Faturas"
                ],
                "summary": "Anula uma fatura",
                "operationId": "c4a8fa6cf9935184ee92c3225e6519ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fatura anulada com sucesso"
                    },
                    "404": {
                        "description": "Fatura não encontrada"
                    }
                }
            }
        },
        "/api/faturas/{id}/pdf": {
            "get": {
                "tags": [
                    "Faturas"
                ],
                "summary": "Gera o PDF da fatura",
                "operationId": "6c735aeba2cf712c23e485101e7e230c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gerador de PDF em desenvolvimento"
                    },
                    "404": {
                        "description": "Fatura não encontrada"
                    }
                }
            }
        },
...
    },
...
}